requirements.txt
pandas
yfinance
matplotlib
numpy
requests
beautifulsoup4
pyyaml